back to decimal

+++ backend/db/migrations/20220705173628-create-spot.js
@@ -56,7 +56,7 @@ module.exports = {
         type: Sequelize.INTEGER,
       },
       avgStarRating: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL,
       },
       previewImage: {
         type: Sequelize.STRING,
